# alora-auth-api/docs/swagger.yaml
basePath: /
definitions:
...
  App:
    properties:
      description:
        type: string
      key:
        example: CRM
        type: string
      launch_url:
        example: https://crm.acme.com/login/initiate?iss=https%3A%2F%2Fcentral.alora.io&target_link_uri=https%3A%2F%2Fcrm.acme.com%2F
        type: string
      name:
        example: Acme CRM
        type: string
      product_id:
        example: 3f7c1b2e-5d4a-4c8b-9e10-2a6b7c8d9e0f
        type: string
      roles:
        example:
        - Editor
        items:
          type: string
        type: array
    type: object
...
  ChangePasswordRequest:
    properties:
      current_password:
        example: correct-horse-battery-staple
        maxLength: 512
        minLength: 1
        type: string
      new_password:
        example: a-different-long-passphrase
        maxLength: 512
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      reqId:
        example: 0b7c0f1e-6a19-4f3f-9a2d-1f9d0f3a6c21
        type: string
    type: object
...
  ManagedGroupRef:
    properties:
      id:
        example: 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      name:
        example: Field sales
        type: string
    type: object
  Me:
    properties:
      authenticated_at:
        type: string
      company:
        $ref: '#/definitions/MeCompany'
      email:
        example: alice@acme.com
        type: string
      id:
        example: 4b5c6d7e-8f90-1a2b-3c4d-5e6f7a8b9c0d
        type: string
      is_admin:
        example: false
        type: boolean
      is_owner:
        example: false
        type: boolean
      manages:
        items:
          $ref: '#/definitions/ManagedGroupRef'
        type: array
      products:
        example:
        - CRM
        - HR
        items:
          type: string
        type: array
      scope_sources:
        items:
          $ref: '#/definitions/ScopeGrant'
        type: array
      scopes:
        example:
        - apps:read
        - users:read
        - users:edit
        items:
          type: string
        type: array
    type: object
  MeCompany:
    properties:
      id:
        example: 5c6d7e8f-9a0b-1c2d-3e4f-5a6b7c8d9e0f
        type: string
      name:
        example: Acme Corp
        type: string
    type: object
...
  ScopeGrant:
    properties:
      group_id:
        example: 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      group_name:
        example: Support
        type: string
      scope:
        example: users:read
        type: string
      source:
        enum:
        - GROUP
        - EXTRA
        example: GROUP
        type: string
    type: object
...
info:
  contact:
    name: Alora Platform
    url: https://alora.io/support
  description: |-
    App Central: one sign-in for every product of the platform, and the OpenID Provider those products trust.

    ## Two kinds of token
    People sign in once at App Central — with a password, Google, or their company's SSO, as their company's login policy allows. That opens an App Central session (an HttpOnly cookie) and yields an App Central access token (audience `app-central`), the ONLY token the `/api` routes accept. It describes everything about App Central: `scope` (what the person may do here) and `products` (the keys of the apps they may open). A product never sees it: when a product is launched, its backend runs an ordinary OpenID Connect authorization-code flow with PKCE against `/oauth/authorize` and `/oauth/token`, and receives a token for itself alone (audience `product:<key>`, carrying only its own roles).

    ## Scopes
...

    ## Group managers
...

    ## Authorizing requests here
    Sign in with `POST /auth/login/password`, copy `access_token` from the response, then press **Authorize** and paste it. Every `/api` route reads the bearer token from the `Authorization` header only — never from a cookie — and re-checks its session against the database on every request.

    ## Companies
...

    ## Errors
...
  license:
    name: Proprietary
  termsOfService: https://alora.io/terms
  title: App Central API
  version: "2.0"
paths:
...
  /api/me:
    get:
      description: Who the caller is, in which company, whether they are an Admin
        or an Owner, and which admin feature keys they hold (an Admin holds all).
        App Central uses it to decide what to show; every route re-checks its own
        guard server-side.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Me'
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Your account
      tags:
      - me
  /api/me/apps:
    get:
      description: The products the caller may use right now — a role through a group
        or granted directly, a live subscription, an active product — each with the
        caller's roles in it and a `launch_url` that starts the product's own sign-in.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/App'
            type: array
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Your apps
      tags:
      - me
  /api/me/change-password:
    post:
      consumes:
      - application/json
      description: 'The current password is re-verified even though the caller holds
        a valid token: that re-authentication is what stops a stolen token from becoming
        permanent account takeover. A wrong one is 403, not 401 — the token was fine,
        and a client that refreshes on 401 must not re-send the guess — and each account
        gets five attempts per fifteen minutes, however many addresses they come from.
        On success every session of the account ends — this one too — so the caller
        signs in again.'
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Password changed; every session ended
        "400":
          description: Malformed body, or the account has no password
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: The current password is incorrect
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Too many attempts for this account
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change your own password
      tags:
      - me
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'An App Central access token from /auth/login/password or /auth/central/refresh.
      Enter it as: Bearer &lt;token&gt;'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...